Reduce stack-allocated array size
This fixes an abort after programming the FPGA that results in an infinite
reset loop of the board.  The call stack depth must be close enough to
the 8K maximum that the 1024-byte array was enough to overflow it, but
a 256-byte array is fine.  I suspect the remainder is being used up
by the SD_MMC calls, but I haven't investigated further.